(#0000) New Type to handle /etc/services entries
This new type "port" handles entries in /etc/services. It uses multiple
key_attributes name and protocol, so you are able to add e.g. multiple
telnet lines for tcp and udp. Sample usage

  port {'telnet':
    number      => '23',
    protocol    => 'tcp',
    description => 'Telnet'
  }

You can also set the protocol in the title. You can use that, to add
entries for udp and tcp since the resourcetitle has to be uniq:

  port {'telnet:udp':
    number      => '23',
    description => 'Telnet'
  }

  port {'telnet:tcp':
    number      => '23',
    description => 'Telnet'
  }

This type is currently not usable because there is no provider and some
hardcoded overrides to be able to test the code