(#0000) Split mounttype - new type mountpoint
Add a new type called mountpoint. This one should only ensure that a
device on a mountpoint is mounted/unmounted. Together with device also
fstype and options can be specified. But they are not required since
there is possibly already an entry in fstab so calling mount without
options and device information will work.

The options are always treated in sync because checking can be very
hard:
* options can have a different name in mount output and what you've
  specified when calling mount. Example. mount -o ro will mount a device
  readonly. On solaris calling mount will now show option1/option2/read
  only/optionN. So we need a translation table ('read only' => 'ro').
* there may be options that show up when calling `mount` that you havent
  specified explicitly. E.g mounting with `mount -o defaults` or without
  any option will become 'rw' when running `mount` afterwards
* you may not be able to see every mountoptions when you run `mount`.
  I think I saw that on HP-UX a while ago: Mount a NFS-Resource with NFS
  specific options (like rsize, wsize) will not show up when running a
  plain mount.