(#7188) Split mounttype - new type mountpoint
Add a new type called mountpoint. This one should only ensure that a
device on a mountpoint is mounted/not mounted.

Valid properties are device, fstype and options but they are not
required because it's likely that there already is an fstab entry for
the described mount.

Mountoptions are currently always treated in sync because checking can
be very hard:
* plain mount command can show different options, e.g. to mount
  read-only you say »mount -o ro« but running a plain mount will report
  options as »read only« on Solaris.
* there may be options that show up when calling `mount` that you havent
  specified explicitly. E.g mounting with `mount -o defaults` or without
  any option will become 'rw' when running `mount` afterwards