(#0000) Split mounttype - new provider mountpoint
Add a provider called mount that uses the output of mount to prefetch
current mounted devices and uses mount/umount to perform changes.

It may not work on any other System than Linux and even on this platform
it is not really well tested so far. Especially prefetching options and
the fstype may fail.