(#0000) Mointpointtype - Better value validation
A mountpoint should always describe an absolute path. It should not
contain spaces because this is also not valid in fstab. Mounting a
device on some mountpath that has a trailing slash may fail. And since
the output of mount will truncate trailing slashes we wont be able to
recognize the mount.

Credits go to John Spray in #6793 that this is fixed now.