(#2538) Refactor execute to not use tempfiles
Change execute to use pipes to directly capture the output of the
childprocess instead of using temporary files the child process writes its
output to.

This allows quite interesting stuff in the future like processing output
while the process is running or tag a process as failed if there was
output on stderr.

Older commits showed that using pipes caused strange blocking issues.
These should not accure when using IO.select.